+++ composer/datasets/lm_datasets.py
@@ -15,7 +15,7 @@ from composer.core.types import Batch
 from composer.datasets.dataloader import DataLoaderHparams
 from composer.datasets.hparams import DatasetHparams
 from composer.utils import dist
-from composer.utils.dynamic_import import MissingConditionalImportError
+from composer.utils.import_helpers import MissingConditionalImportError
 
 __all__ = ["LMDatasetHparams"]
