+++ composer/loggers/logger_hparams.py
@@ -259,7 +259,7 @@ class ObjectStoreLoggerHparams(LoggerDestinationHparams):
             The artifact name will be a string. The function should return a boolean indicating whether the artifact
             should be logged.
 
-            .. seealso: :func:`composer.utils.dynamic_import.import_object`
+            .. seealso: :func:`composer.utils.import_helpers.import_object`
 
             Setting this parameter to ``None`` (the default) will log all artifacts.
         object_name_format (str, optional): See :class:`~composer.loggers.object_store_logger.ObjectStoreLogger`.
