+++ composer/models/gpt2/gpt2_hparams.py
@@ -6,7 +6,7 @@ import dataclasses
 from typing import TYPE_CHECKING
 
 from composer.models.transformer_hparams import TransformerHparams
-from composer.utils.dynamic_import import MissingConditionalImportError
+from composer.utils.import_helpers import MissingConditionalImportError
 
 if TYPE_CHECKING:
     from composer.models.transformer_shared import ComposerTransformer
