+++ composer/utils/__init__.py
@@ -2,8 +2,8 @@
 
 """Helper utilities."""
 from composer.utils.checkpoint import load_checkpoint, save_checkpoint
-from composer.utils.dynamic_import import import_object
 from composer.utils.file_retriever import get_file
+from composer.utils.import_helpers import MissingConditionalImportError, import_object
 from composer.utils.iter_helpers import ensure_tuple, iterate_with_pbar, map_collection
 from composer.utils.object_store import ObjectStore, ObjectStoreHparams
 from composer.utils.string_enum import StringEnum
